[#71043] 글 삭제 비즈니스 로직 service로 변경 및 쿼리 수정
* 게시글 삭제시 이미지도 삭제하는 부분 service로 이동
* 쿼리문 위에 @Query를 붙이지 않고도 쿼리메소드로 가능하므로 제거 및 쿼리메소드로 변경
* 이전의 게시글 삭제 메소드 제거

+++ src/main/java/com/zum/pilot/service/impl/PostServiceImpl.java
@@ -82,7 +82,6 @@ public class PostServiceImpl implements PostService {
   @Override
   @Transactional
   public void modifyPost(Long postId, String title, String content, MultipartFile file) {
-//    PostEntity postEntity = postService.getPost(postId);
     PostEntity postEntity = postRepository.findByIdAndDeleteFlag(postId, false);
     postEntity.setTitle(title);
     postEntity.setContent(content);
@@ -114,10 +113,19 @@ public class PostServiceImpl implements PostService {
 
   @Override
   @Transactional
-  public void deletePost(Long postId) {
+  public void deletePost(Long postId, Long authId) {
     PostEntity postEntity = postRepository.findByIdAndDeleteFlag(postId, false);
-    postEntity.setDeleteFlag(true);
+    if (postEntity.getUserId() != authId) {
+      return;
+    }
+    // 이미지가 있다면 삭제
+    if (!"".equals(postEntity.getImagePath())) {
+      File uploadFile = new File(uploadPath + "/" + postEntity.getImagePath());
+      if (uploadFile.exists() && uploadFile.isFile())
+        uploadFile.delete();
+    }
     // 게시글 삭제
+    postEntity.setDeleteFlag(true);
     postRepository.save(postEntity);
     commentService.deleteCommentByPostId(postId);
   }
